Replace deprecated Filter interface with FilterLabel
Use physical filter names for jointcal's internal CcdImage structure to
distinguish e.g. HSC-R and HSC-R2 if we ever support a multi-filter mode.

Change "filter" to "band" throughout the jointcal.py to make the type of label
more clear. In this case it is used to load refcats, which expect a band.